pyproject.toml
[project]
name = "macllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bashlex>=0.18",
    "beautifulsoup4>=4.12.3",
    "certifi>=2024.12.14",
    "litellm>=1.0.0",
    "maccal>=1.0.0",
    "markdown-it-py>=4.0.0",
    "openai[realtime]>=1.54.4",
    "pyobjc>=10.3.1",
    "pytest>=8.3.5",
    "quickmachotkey>=2023.11.17",
    "requests>=2.32.3",
    "smolagents>=0.1.0",
    "things.py>=1.0.1",
    "tomli_w>=1.0.0",
    "txtai>=9.3.0",
]


[tool.pytest.ini_options]
addopts = "-m 'not external and not calendar and not things'"
markers = [
    "external: tests that hit external services",
    "calendar: tests that access macOS calendar (requires permissions)",
    "things: tests that access the real Things app and database (requires Things URLs auth)",
    "sandbox: tests that exercise macOS sandbox-exec (requires macOS)",
]
filterwarnings = [
    "ignore::UserWarning:pydantic.*",
    "ignore:There is no current event loop:DeprecationWarning:litellm.*",
]
